frontend/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
cryptography
stegano
python-dotenv
transformers
